prospect-outreach/requirements.txt
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
twilio>=8.0.0
sib-api-v3-sdk>=7.0.0
reportlab>=4.0.0
Pillow>=10.0.0
requests>=2.28.0
python-dotenv>=1.0.0
jinja2>=3.1.0
